Qt 6.8.1 (17)

+++ configure.sh
@@ -150,7 +150,11 @@ elif [ $1 = "linux" ]; then
     cp -r "$path"/vlc bin
 
     cp "$path"/libvlc*.so* bin
-    cp "$path"/libidn.so*  bin
+
+    if [ -f "$path"/libidn.so* ]; then
+
+        cp "$path"/libidn.so* bin
+    fi
 fi
 
 #--------------------------------------------------------------------------------------------------
